Send vscode client configuration during initialization
